finalizing changes for veraweb specific release, updated pom dependencies
the actual changes included are are wrapper mechanism using dynamic proxies that will circumvent a
possible configuration problem with postgresql where there are timestamps with timezone information
stored in the database that have a different timezone than the postgresql server and the webserver
accessing the dbms. in that case, whenever a timestamp was instantiated from the database, the time
would shift everytime that the data was accessed.