w.i.p.: rewrite delegation stuff from scratch
Moved stuff around quite a bit. Delegation form was redone from scratch.
(yet to do: image loading and resizing works, but they are not send to
the backend yet.)
It is the most complex component on the client side and will hopefully
serve as an example of how to structure things in the future.