w.i.p. sanitizing backend requests / use bootstrap date picker
it does make no sense at all to include the osiam user id
in the uri path if actually we are only allowed to see/modify
data for our own account. This information is already included in
the hmac token we send with each request via cookie header.
Fixed for contact form and a few places.

further, switched to bootstrap date picker (same as in FES branch).