work around the fix for a CVE breaking everything
Tomcat CVE-2020-1938 “Ghostcat (AJP)” introduced a new Connector
attribute “secretRequired” *and* defaults it to "true", despite
old (as in, anything in a released version/branch; it is only
present in an unstable development snapshot targetting Apache 2.6
and not backported to 2.4) Apache mod_jk versions not even having
a function to set a secret. As our setup only allows connections
from localhost and we recommend a single-purpose system, we must
disable it.

Note that we’re guesstimating the 8.0 version, given upstream
doesn’t support 8.0 at all any more (only 7.0 and 8.5 and of
course 9.0 onwards), but Debian jessie LTS does.